tools/dpa/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dpa-server"
version = "0.1.0"
description = "Deep Potential (DP) model training and calculation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ase==3.26.0",
    "bohr-agent-sdk>=0.1.101",
    "numpy==2.2.0",
    "dpdata>=0.2.25",
    "deepmd-kit==3.1.0",
    "pydflow>=1.8.132"
]

[tool.setuptools.packages.find]
where = ["src"]
